// src/models/RemoteAnnouncementDetail.ts
export type RemoteAnnouncementDetail = {
    boardIdx: string;
    boardSubIdx: number;
    notice: string;
    memberType: string;
    memberIdx: string;
    memberEmail: string;
    memberName: string;
    hit: number;
    useTag: string;
    createdAt: string;
    boardSubject: string;
    boardContents: string;
    reply: string[];
};